Fixing warning and default preview compile immediately

+++ source/Return/vertex_components.h
@@ -15,5 +15,5 @@ namespace gfx
     const char* const g_vertex_component_names[] = { "Float", "Bool", "Int", "Vec2", "Vec3" };
 
     int component_size(VertexComponent);
-    int vertex_size(const VertexComponent*, int n);
+    int vertex_size(const VertexComponent*, size_t n);
 }
